+++ express-zod-api/tests/endpoint.spec.ts
@@ -282,7 +282,9 @@ describe("Endpoint", () => {
           output: z.object({ something: z.number() }),
           handler: vi.fn(),
         });
-        const responses = endpoint.getResponses(variant);
+        const responses = endpoint.getResponses(variant, {
+          maxCombinations: 0,
+        });
         expect(responses).toMatchSnapshot();
         expect(() => (responses as any[]).push()).toThrowError(/read only/);
       },